requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.28.69
botocore==1.31.69
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distro==1.8.0
Flask==2.2.3
frozenlist==1.3.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
openai==1.3.6
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
typing_extensions==4.8.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
